Move scan and scanMerge to stream module
