--- stream/scan.js
@@ -1,13 +1,0 @@
-"use strict"
-
-var combine = require("./stream").combine
-
-module.exports = function (reducer, seed, stream) {
-	var newStream = combine(function (s) {
-		return seed = reducer(seed, s._state.value)
-	}, [stream])
-
-	if (newStream._state.state === 0) newStream(seed)
-
-	return newStream
-}
